minor refactor for code reuse
move some binning methods out of cam_touptek.cpp into image_math.h
so the methods can be used by another camera in a forthcoming PR

no functional changes
